Move /schedule route out of base, use get_argument, remove post = get
